change codes for new api;change dtype of dataset as fp32

+++ spikingjelly/datasets/__init__.py
@@ -182,7 +182,7 @@ def load_npz_frames(file_name: str) -> np.ndarray:
     :return: frames
     :rtype: np.ndarray
     '''
-    return np.load(file_name, allow_pickle=True)['frames']
+    return np.load(file_name, allow_pickle=True)['frames'].astype(np.float32)
 
 def integrate_events_segment_to_frame(x: np.ndarray, y: np.ndarray, p: np.ndarray, H: int, W: int, j_l: int = 0, j_r: int = -1) -> np.ndarray:
     '''
